Move test helpers into libs.tests
